fix(stella-tui): one word wrap in the crate, and it keeps a line break
Three private greedy wrappers with the same job:
`views/cards.rs::wrap` (shared by the plan card and the task zoom since
#5041), `views/mcp_tab/detail.rs::wrap`, and `views/models_card.rs::wrap_parts`.

**They were not equivalent, which is what made this worth doing rather than
a tidy-up.** The greedy packing agreed, but the two card copies flattened
their input with `split_whitespace`, so a two-paragraph server description
came back as one run-on line, while `mcp_tab`'s wrapped per line and guarded
`width == 0`. The shared one now takes the superset — a line break in the
input is a line break in the output, a blank line survives as a blank row,
and zero columns wraps to nothing — because that is what a reader of the
text wants and flattening is a loss with nothing to recommend it.

`mcp_tab/detail.rs`'s copy is deleted and its callers take the shared one.
Its test came with it: `wrap_breaks_on_words_and_never_loses_one` was the
only test any of the three had, which is the half of the duplication that
cost the most — the copies were not equally proved.
`a_paragraph_break_survives_the_wrap` is the new one, pinning the behaviour
the merge chose.

`wrap_parts` stays, documented as the deliberate third case #5156 allows for
it: it packs pre-split parts rather than breaking on whitespace, joins with
` · `, and truncates a part too wide for a row where `wrap` gives it a row
of its own. Folding it in would have made every one of those a special case
of a function that has no reason to know about them.

**No golden moved**, which is what the issue asks for — the full
`stella-tui` suite passes untouched, so nothing in the deck's fixtures wraps
text containing a newline and the behaviour change is confined to inputs
that previously lost information.

Remaining work filed rather than folded in: `wrap` counts `char`s where it
needs display columns, so wide text overflows the card it was measured for.
All three copies did that, and fixing it moves goldens for a reason
unrelated to this merge — #5307.

Closes #5156
